Make use submodule optional
This introduces a configuration switch `use_submodules` that
controls whether external dependencies are added as submodules
or whether they are searched for by CMake.

Additionally, the deploy tests are switched to using no submodules.
The idea behind that is that the setup with submodules is covered
quite well in the local tests, where as the no-submodule case requires
adaptation of the CI workflow.

This currently affects Catch2 and pybind11.